services/dashboard/requirements.txt
streamlit==1.31.0
confluent-kafka==2.3.0
psycopg2-binary==2.9.9
pandas>=2.0.0
numpy>=1.24.0,<2.0.0
opencv-python-headless==4.9.0.80
plotly>=5.18.0
